Reorganize functions.
The globally unique identifiers are not unique to the blocked URI
functions. Let's break them out into a separate file to be able to be
used via other modules.
