+++ lib/sanitize/blockedURI.js
@@ -1,59 +1,10 @@
 'use strict';
 
+var getGloballyUniqueIdentifiers = require('../util').getGloballyUniqueIdentifiers;
+var isGloballyUniqueIdentifier = require('../util').isGloballyUniqueIdentifier;
+var sanitizeGloballyUniqueIdentifier = require('../util').sanitizeGloballyUniqueIdentifier;
 var url = require('url');
 var validator = require('validator');
-
-/**
- * List of allowed globally unique identifiers.
- *
- * @return {String}    Sanitized globally unique identifier.
- */
-function getGloballyUniqueIdentifiers() {
-  return [
-    'data',
-    'filesystem',
-    'blob'
-  ];
-}
-
-/**
- * Determines if the value uses a globally unique identifier.
- *
- * @param  {String}    value    The raw `blocked-uri` value.
- * @return {String}             Boolean representing whether or not the value is
- *                              a globally unique identifier.
- */
-function isGloballyUniqueIdentifier(value) {
-  var globallyUniqueIdentifiers = getGloballyUniqueIdentifiers();
-  var result = false;
-
-  globallyUniqueIdentifiers.forEach(function(globallyUniqueIdentifier) {
-    if (0 === value.indexOf(globallyUniqueIdentifier)) {
-      result = true;
-    }
-  });
-
-  return result;
-}
-
-/**
- * Sanitizes a globally unique identifier `blocked-uri` value.
- *
- * @param  {String}    value    The raw `blocked-uri` value.
- * @return {String}             Sanitized globally unique identifier.
- */
-function sanitizeGloballyUniqueIdentifier(value) {
-  var globallyUniqueIdentifiers = getGloballyUniqueIdentifiers();
-  var cleanValue = '';
-
-  globallyUniqueIdentifiers.forEach(function(globallyUniqueIdentifier) {
-    if (0 === value.indexOf(globallyUniqueIdentifier)) {
-      cleanValue = globallyUniqueIdentifier;
-    }
-  });
-
-  return cleanValue;
-}
 
 /**
  * Sanitizes a `blocked-uri` value.
